fix/v0-shims: Export VideoProps, split storybooks
