+++ packages/react-components/react-migration-v0-v9/library/src/index.ts
@@ -4,6 +4,7 @@ export { FormFieldShim } from './components/FormField';
 export { Segment } from './components/Segment';
 export { slider } from './components/Slider';
 export { Video, videoClassName } from './components/Video';
+export { type VideoProps } from './components/Video';
 export { input } from './components/Input';
 export { v0Icon, v9CustomSizeIcon, v9DisabledCursor, v9HoverClasses, v9Icon } from './components/Button';
 export { spinner } from './components/Spinner';
