+++ packages/react-components/react-migration-v0-v9/stories/src/Attachment/Default.stories.tsx
@@ -1,0 +1,13 @@
+import { Attachment, AttachmentHeader, AttachmentBody } from '@fluentui/react-migration-v0-v9';
+
+export const DefaultAttachment = () => {
+  return (
+    <Attachment>
+      <AttachmentBody>
+        <AttachmentHeader>Document.docx</AttachmentHeader>
+      </AttachmentBody>
+    </Attachment>
+  );
+};
+
+DefaultAttachment.storyName = 'Default';
