+++ packages/react-components/react-migration-v0-v9/stories/src/Attachment/Header.stories.tsx
@@ -1,0 +1,14 @@
+import * as React from 'react';
+import { Attachment, AttachmentHeader, AttachmentBody } from '@fluentui/react-migration-v0-v9';
+
+export const Header = () => {
+  return (
+    <Attachment>
+      <AttachmentBody>
+        <AttachmentHeader>Strategy.docx</AttachmentHeader>
+      </AttachmentBody>
+    </Attachment>
+  );
+};
+
+Header.storyName = 'Header';
