+++ packages/react-components/react-migration-v0-v9/stories/src/Video/AutoPlay.stories.tsx
@@ -1,0 +1,7 @@
+import * as React from 'react';
+import { Video } from '@fluentui/react-migration-v0-v9';
+
+export const AutoPlay = () => (
+  <Video src="https://fabricweb.azureedge.net/fabric-website/assets/videos/2020_MSFT_Icon_Celebration.mp4" autoPlay />
+);
+AutoPlay.storyName = 'AutoPlay';
