+++ packages/react-components/react-migration-v0-v9/stories/src/Video/Muted.stories.tsx
@@ -1,0 +1,7 @@
+import * as React from 'react';
+import { Video } from '@fluentui/react-migration-v0-v9';
+
+export const Muted = () => (
+  <Video src="https://fabricweb.azureedge.net/fabric-website/assets/videos/2020_MSFT_Icon_Celebration.mp4" muted />
+);
+Muted.storyName = 'Muted';
