+++ packages/react-components/react-migration-v0-v9/stories/src/Video/index.stories.tsx
@@ -3,29 +3,10 @@ import { Video } from '@fluentui/react-migration-v0-v9';
 
 import descriptionMd from './VideoDescription.md';
 
-export const Default = () => (
-  <Video src="https://fabricweb.azureedge.net/fabric-website/assets/videos/2020_MSFT_Icon_Celebration.mp4" controls />
-);
-Default.storyName = 'Default';
-
-export const AutoPlay = () => (
-  <Video src="https://fabricweb.azureedge.net/fabric-website/assets/videos/2020_MSFT_Icon_Celebration.mp4" autoPlay />
-);
-AutoPlay.storyName = 'AutoPlay';
-
-export const Muted = () => (
-  <Video src="https://fabricweb.azureedge.net/fabric-website/assets/videos/2020_MSFT_Icon_Celebration.mp4" muted />
-);
-Muted.storyName = 'Muted';
-
-export const WithPoster = () => (
-  <Video
-    src="https://fabricweb.azureedge.net/fabric-website/assets/videos/2020_MSFT_Icon_Celebration.mp4"
-    poster="https://fabricweb.azureedge.net/fabric-website/assets/images/avatar/large/ade.jpg"
-    controls
-  />
-);
-WithPoster.storyName = 'WithPoster';
+export { Default as DefaultVideo } from './Default.stories';
+export { AutoPlay } from './AutoPlay.stories';
+export { Muted } from './Muted.stories';
+export { WithPoster } from './WithPoster.stories';
 
 export default {
   title: 'Migration Shims/V0/Video',
